feat(main): add context option to commit message
This commit adds a new option `-c, --context <context>` to the `commit` command, which allows users to add context or instructions to their commit message. The context will be added as a new message in the OpenAI API request, and will be included in the final commit message if the user chooses to confirm it.

This change also updates the `request` function to accept an array of messages, which allows the context message to be added to the OpenAI API request. The final commit message is now generated from the last message in the array, which should always be the message generated by the OpenAI API.

Lastly, this commit removes the `set-mode` command, which was previously used to select a commit style. This feature is no longer necessary, as the commit message format is now fixed and generated by the OpenAI API.