fix(commit): enable resetting changes
Support resetting changes in case the user wants to cancel the commit process. This is done by listening to the SIGINT, SIGTERM, and SIGHUP signals and calling the `git reset` command. Additionally, if the user presses `Ctrl+C`, the same reset function is called.

No issues are fixed by this commit.