feat(commit): add chunking feature for large changes
This commit adds a new feature to chunk the files into smaller sets when the changes are too big to generate a commit message. The user will be prompted to confirm whether they want to chunk the files or not. If they choose to chunk, the program will use GPT-3 to determine the scope of changes from the git status and chunk the files accordingly. The commit message will be generated for each chunk separately.